pyproject.toml
[tool.poetry]
name = "easy_bookmark"
version = "0.1.0"
description = "Redefine your bookmarks"
authors = ["CarlosVecina <carlosvecinatebar@gmail.com>", "Ana Barbero <anabarberocastejon@gmail.com>"]
readme = "README.md"

packages = [{include = "pic2bookmark", from = "."}]

[tool.poetry.dependencies]
python = "~3.11"
pandas = "^2.2.2"
easyocr = "^1.7.1"
openai = "^1.35.10"
python-dotenv = "^1.0.1"
python-telegram-bot = "^21.3"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
